Update package versions and refactor code
