+++ src/FeatureManagement.Database.CosmosDB/FeatureManagement.Database.CosmosDB.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.1;net6.0;net8.0</TargetFrameworks>
+    <AzureCosmosDisableNewtonsoftJsonCheck>true</AzureCosmosDisableNewtonsoftJsonCheck>
   </PropertyGroup>
 
   <Import Project="..\..\build\NuGet.props" />
