+++ tests/FeatureManagement.Database.EntityFrameworkCore.Tests/FeatureManagement.Database.EntityFrameworkCore.Tests.csproj
@@ -36,7 +36,7 @@
   </ItemGroup>
 
   <ItemGroup Condition="'$(TargetFramework)' == 'net6.0'">
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" VersionOverride="6.0.32" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" VersionOverride="6.0.36" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Design" VersionOverride="7.0.20" />
     <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" VersionOverride="7.0.20">
       <PrivateAssets>all</PrivateAssets>
